feat: session을 free와 paid로 구분
