+++ README.md
@@ -17,7 +17,7 @@
   * [X] 이미지 크기는 1MB 이하여야 한다.
   * [X] 이미지 타입은 gif, jpg(jpeg 포함), png, svg만 허용한다.
   * [X] 이미지의 width는 300픽셀, height는 200픽셀 이상이어야 하며, width와 height의 비율은 3:2여야 한다.
-* [ ] 강의는 무료 강의와 유료 강의로 나뉜다.
+* [X] 강의는 무료 강의와 유료 강의로 나뉜다.
   * [ ] 무료 강의는 최대 수강 인원 제한이 없다.
   * [ ] 유료 강의는 강의 최대 수강 인원을 초과할 수 없다.
   * [ ] 유료 강의는 수강생이 결제한 금액과 수강료가 일치할 때 수강 신청이 가능하다.
